Tests/LibWeb/Text/expected/HTML/HTMLInputElement-cloning-steps.txt

Cloned checkbox checked: true
Cloned text input value: PASS
